+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * [#1638](https://github.com/bbatsov/rubocop/issues/1638): Use Parser functionality rather than regular expressions for matching comments in `FirstParameterIndentation`. ([@jonas054][])
 * [#1642](https://github.com/bbatsov/rubocop/issues/1642): Raise the correct exception if the configuration file is malformed. ([@bquorning][])
+* [#1647](https://github.com/bbatsov/rubocop/issues/1647): Skip `SpaceAroundBlockParameters` when lambda has no argument. ([@eitoball][])
 
 ## 0.29.0 (05/02/2015)
 
@@ -1254,3 +1255,4 @@
 [@mmozuras]: https://github.com/mmozuras
 [@d4rk5eed]: https://github.com/d4rk5eed
 [@cshaffer]: https://github.com/cshaffer
+[@eitoball]: https://github.com/eitoball